template: share VM template building
 - deprecate template attribute
 - replace with detailed attributes common with VM
 - new schema part shared: tags
 - change os and graphics schemas to TypeList and add tests
