+++ opennebula/resource_opennebula_virtual_machine_test.go
@@ -29,6 +29,13 @@ func TestAccVirtualMachine(t *testing.T) {
 					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "permissions", "642"),
 					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "memory", "128"),
 					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "cpu", "0.1"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "graphics.#", "1"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "graphics.0.keymap", "en-us"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "graphics.0.listen", "0.0.0.0"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "graphics.0.type", "VNC"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "os.#", "1"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "os.0.arch", "x86_64"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "os.0.boot", ""),
 					resource.TestCheckResourceAttrSet("opennebula_virtual_machine.test", "uid"),
 					resource.TestCheckResourceAttrSet("opennebula_virtual_machine.test", "gid"),
 					resource.TestCheckResourceAttrSet("opennebula_virtual_machine.test", "uname"),
@@ -50,6 +57,13 @@ func TestAccVirtualMachine(t *testing.T) {
 					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "group", "oneadmin"),
 					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "memory", "196"),
 					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "cpu", "0.2"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "graphics.#", "1"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "graphics.0.keymap", "en-us"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "graphics.0.listen", "0.0.0.0"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "graphics.0.type", "VNC"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "os.#", "1"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "os.0.arch", "x86_64"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "os.0.boot", ""),
 					resource.TestCheckResourceAttrSet("opennebula_virtual_machine.test", "uid"),
 					resource.TestCheckResourceAttrSet("opennebula_virtual_machine.test", "gid"),
 					resource.TestCheckResourceAttrSet("opennebula_virtual_machine.test", "uname"),
